cmp-cnm/requirements.txt
django<3.3
django-filter
django-netfields
djangorestframework<3.14
drf_yasg
gunicorn
httptools
psycopg2-binary
python-dotenv
requests
uvicorn
uvloop
openstacksdk
django-apscheduler
ncclient

oslo.config>=6.0.0 # Apache-2.0
oslo.messaging>=8.0.0 # Apache-2.0
